% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baci.R
\name{baci}
\alias{baci}
\title{Compute bias adjusted confidence interval using truncated exponential prior distributions for kD, kY.}
\usage{
baci(fit, treatment, benchmark, N = 1000, alpha = 5/100, medkd = 1, medky = 1)
}
\arguments{
\item{fit}{An object of class \code{lm}.}

\item{treatment}{A character string naming the treatment variable.}

\item{benchmark}{A character string naming the benchmark variable.}

\item{N}{Numeric value for grid size.}

\item{alpha}{Significance level.}

\item{medkd}{Median of the distribution of kD (default=1).}

\item{medky}{Median of the distribution of kY (default=1).}
}
\value{
A list containing the following two elements:
\item{results}{A data frame with 2 rows ("Unadjusted" and "Bias-adjusted") and
2 columns ("Lower" and "Upper") containing the computed 100*(1-alpha)\\%
unadjusted and bias-adjusted confidence intervals.}
\item{undstats}{A data frame of underlying statistics: the estimate, std error,
max(kD) and max(k(Y).}
\item{support_kdky_plot}{A \code{ggplot2} plot object visualizing the support of
the joint distribution of (kD,kY).}
}
\description{
Compute bias adjusted confidence interval using truncated exponential prior distributions for kD, kY.
}
\examples{
# CRAN requires checking if the optional package is installed before running the example
if (requireNamespace("sensemakr", quietly = TRUE)) {
  library(sensemakr)

  # Define example variables
  myN <- 500

  # Fit the model using the darfur dataset provided by sensemakr
  fit <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
              pastvoted + hhsize_darfur + female + village,
            data = sensemakr::darfur)

  # Run baci function and see results
  res1 <- baci(fit = fit, treatment = "directlyharmed",
               benchmark = "female", N = myN, alpha = 1/100)
}
}
