% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtruncexp.R
\name{dtruncexp}
\alias{dtruncexp}
\title{truncated exponential density}
\usage{
dtruncexp(x, rate = 1, a = 0, b = Inf)
}
\arguments{
\item{x}{point between a and b at which density is computed}

\item{rate}{rate parameter}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
a real number
}
\description{
truncated exponential density
}
\examples{
dtruncexp(x=1, rate=1, a=0, b=10)
}
