% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtruncgamma.R
\name{dtruncgamma}
\alias{dtruncgamma}
\title{truncated gamma density}
\usage{
dtruncgamma(x, shape, scale, a = 0, b = Inf)
}
\arguments{
\item{x}{point at which density is computed}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{a}{lower limit (default is 0)}

\item{b}{upper limit (default is Inf)}
}
\value{
a real number
}
\description{
truncated gamma density
}
\examples{
dtruncgamma(x=1,shape=2, scale=3, a=0, b=20)
}
