% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linvx.R
\name{linvx}
\alias{linvx}
\title{quasi-triangular probability distribution function}
\usage{
linvx(x, xvec, k)
}
\arguments{
\item{x}{point (scalar) at which pdf is evaluated}

\item{xvec}{vector of all possible x values}

\item{k}{mode and median of the distribution}
}
\value{
the value (scalar) of the pdf at x
}
\description{
\code{linvx()} is deprecated; please use \code{\link[=baci]{baci()}} instead.
}
\examples{
xfull <- runif(n=100,min=0,max=10)
xpoint <- 5
xmod <- 2
res_pdf <- linvx(x=xpoint,xvec=xfull,k=xmod)

}
\keyword{internal}
